vcs_report: added actor generating report about VCS
Generate input data for report actor. The actor uses tag
"informational" as these data will not affect transition to the
new system. This is completele independent on the generated status
for the report actor.

In future, actors like this one could be skipped based on the
informational tag with all other actors, that generates output used
only by "informational" actors.